WordCamp Site Cloner: Refactor for compatibility with WP 4.7 and Jetpack 4.2.2
WordPress 4.7 added a Customizer feature that allows editing CSS. Jetpack 4.2.2 replaced the Custom CSS module with an entirely new one that is compatible with the Customizer feature, but broke backwards-compatibility with the Site Cloner and Remote CSS plugins.

This commit refactors the Site Cloner to restore compatibility with Jetpack, and add support for the new Core feature.

https://make.wordpress.org/core/2016/10/11/feature-proposal-better-theme-customizations-via-custom-css-with-live-previews/
https://make.wordpress.org/core/2016/11/26/extending-the-custom-css-editor/
https://github.com/Automattic/jetpack/issues/5539
https://github.com/Automattic/jetpack/pull/5804
https://core.trac.wordpress.org/ticket/35395
https://core.trac.wordpress.org/ticket/38672
https://core.trac.wordpress.org/ticket/39270

git-svn-id: https://meta.svn.wordpress.org/sites/trunk/wordcamp.org/public_html/wp-content/plugins/wcpt@4577 74240141-8908-4e6f-9713-ba540dce6ec7