WordCamp: Limit metabox_save action to WCPT post type
WCPT's `metabox_save` method was hooked to `save_post`, so it was firing
for every save, even though it's intended only for the WCPT, as
evidenced by the conditional return within the method if it's not the
right post type.

The problem is that it was throwing errors when it got called during
unrelated `wp_insert_post` invocations. The workaround used in
`create_post_stubs` was to temporarily remove the action from the hook
and then add it back. However with the recent addition of the
`wcpt_configure_new_site` hook, external plugins can run into the same
errors, and forcing them to also temporarily remove the action is
smelly.

The solution in this commit is to stop using the `save_post` hook, and
instead use `save_post_{$post->post_type}` to ensure that the
`metabox_save` method only fires on the post type for which it is
intended.

git-svn-id: https://meta.svn.wordpress.org/sites/trunk/wordcamp.org/public_html/wp-content/plugins/wcpt@5365 74240141-8908-4e6f-9713-ba540dce6ec7