WordCamp: Use `_status_change_log_%' to filter WordCamp Applications.
On WordCamp Application Status report page, we were fetching 300 WordCamp posts, iterating over them one by one and figuring out if we should show in the report.
Changed this to directly filter in database for the posts that we want to show. This allows us to limit the 300 limit as we are already querying in the Database. 

Fixes #3911



git-svn-id: https://meta.svn.wordpress.org/sites/trunk/wordcamp.org/public_html/wp-content/plugins/wcpt@7841 74240141-8908-4e6f-9713-ba540dce6ec7